Correct package hierarchy
In Python 3, implicit access to members of the same package doesn't behave like Python 2.
Hide the complexity of the package hierarchy by having its main elements available from the root.
Changes lead to circular imports. As classes are small, correct the problem by using only one module for elements.
Name of the module is no longer important as it was corrected in AllocineObject.__new__ in last commit.
Extract the test code to a new separate file.